Fixup minor ui bugs
