+++ package.json
@@ -77,7 +77,7 @@
     "@appnest/masonry-layout": "^2.2.3",
     "@glidejs/glide": "^3.6.1",
     "@stackla/handlebars": "^0.1.3",
-    "@stackla/widget-utils": "^0.0.0-20241107205544",
+    "@stackla/widget-utils": "^0.0.0-20241108030542",
     "cookie-parser": "^1.4.6",
     "happy-dom": "^14.12.3",
     "hbs": "^4.2.0",
