+++ widgets/storypage/widget.scss
@@ -6,7 +6,7 @@
 
 .grid {
   display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
+  grid-template-columns: 1fr 1fr 1fr 1fr 1fr;
 
   @include styles.breakpoint(tablet) {
     grid-template-columns: 1fr 1fr 1fr;
