+++ widgets/storypage/widget.tsx
@@ -2,7 +2,7 @@ import { loadAllUnloadedTiles } from "@stackla/widget-utils/extensions/swiper"
 import { loadWidget } from "@stackla/widget-utils"
 import shopspotStyle from "./components/shopspot-icon/base.scss"
 import "./direct-uploader.component"
-import { loadDirectUploaderListeners, loadDirectUploaderTileButton } from "./direct-uploader.lib"
+import { loadDirectUploaderTileButton } from "./direct-uploader.lib"
 
 loadWidget({
   extensions: {},
@@ -30,5 +30,5 @@ loadWidget({
     }
   }
 })
+
 loadAllUnloadedTiles()
-loadDirectUploaderListeners()
